Transparent backgrounds, misc improvements, and some NPC groundwork

+++ src/World/Entity/GlowingEntity.java
@@ -1,5 +1,5 @@
 package World.Entity;
 
 public interface GlowingEntity {
-    public int glowRadius();
+    public double glowRadius();
 }
